Add static method to setup (~ git init) tracking folders and commit_head_pointer.json Fix checkout function to start with the correct diff in backward traversal Fix checkout function to stop at the correct diff file depending on forward/backward traversal Add "direction" argument to checkout to validate checkpoints easily

+++ difftools.py
@@ -1,7 +1,5 @@
-import onnx
 import torch
 import numpy as np
-from onnx import numpy_helper as nh
 
 import copy
 import pickle
@@ -53,13 +51,14 @@ class DiffTools:
         self.commit_head_info = CommitHeadInfo()
 
     @staticmethod
-    def setup(model_name=None, reset_folders=False):
+    def setup(base_model_name=None, reset_folders=False):
         if (reset_folders):
             print(f"Resetting diff tracking in {os.getcwd()} to a clean slate.")
             # Get path to model head (onnx file)
             commit_head_info = CommitHeadInfo()
-            print(f"Deleting model head -: {commit_head_info.global_head_model_path}")
-            os.remove(commit_head_info.global_head_model_path)
+            if(os.path.exists(commit_head_info.global_head_model_path)):
+                print(f"Deleting model head -: {commit_head_info.global_head_model_path}")
+                os.remove(commit_head_info.global_head_model_path)
 
             if (os.path.exists(constants.COMMITS_FOLDER)):
                 for file in sorted(os.listdir(constants.COMMITS_FOLDER)):
@@ -70,11 +69,11 @@ class DiffTools:
 
         os.makedirs(constants.COMMITS_FOLDER, exist_ok=True)
         # Append _head for model head name
-        model_head_name = model_name.split(".")[0] + "_head" + model_name.split(".")[1]
+        model_head_name = base_model_name.split(".")[0] + "_head" + "." + base_model_name.split(".")[1]
         commit_head_pointer = {
             "current_head": "",
             "global_head_model": os.path.join(constants.MODELS_FOLDER, model_head_name),
-            "global_tail_model": os.path.join(constants.MODELS_FOLDER, model_name),
+            "global_tail_model": os.path.join(constants.MODELS_FOLDER, base_model_name),
             "global_head": "",
             "global_tail": ""
         }
@@ -91,7 +90,7 @@ class DiffTools:
         # Fetch Global Head's diff index value
         if(self.commit_head_info.global_head):
             print("Loading Global Head File -: {}".format(self.commit_head_info.global_head))
-            with open("{}.modeldiff".format(self.commit_head_info.global_head), "rb") as infile:
+            with open(self.commit_head_info.global_head, "rb") as infile:
                 global_head_diff = pickle.load(infile)
 
         # Each entry in changes should be a 3-tuple (initializer, update_type, delta)
@@ -103,11 +102,11 @@ class DiffTools:
                 update.initializer_name: (update.update_type, update.delta) for update in updates
             }
         }
-        print("New diff chain link (index, prev, next) -:", update_dict["diff_index"], update_dict["previous_diff"], update_dict["next_diff"])
+        print("New diff chain link (index, prev diff, next diff) -:", update_dict["diff_index"], update_dict["previous_diff"], update_dict["next_diff"])
 
         hash = hashlib.md5(pickle.dumps(update_dict)).hexdigest()
         new_diff_name = os.path.join(constants.COMMITS_FOLDER, f"{hash}.modeldiff")
-        print(new_diff_name)
+        print(f"New diff name -: {new_diff_name}")
         # Add name of new diff file to update_dict for easier access
         update_dict["current_diff"] = new_diff_name
 
@@ -130,19 +129,24 @@ class DiffTools:
 
         # Apply new diff to model and persist it at head
         print("Persisting new model")
-        model_head = self.apply_diff_file(AugumentedOnnxModel(self.commit_head_info.global_head_model_path), self.commit_head_info.global_head)
+        if(update_dict['diff_index'] == 1):
+            # First diff, model head doesn't exist yet. Hence use model tail
+            prev_model_head = AugumentedOnnxModel(self.commit_head_info.global_tail_model_path)
+        else:
+            prev_model_head = AugumentedOnnxModel(self.commit_head_info.global_head_model_path)
+        model_head = self.apply_diff_file(prev_model_head, self.commit_head_info.global_head)
         model_head.save_model(self.commit_head_info.global_head_model_path)
 
         return new_diff_name
 
 
-    def apply_diff_file(self, augumented_onnx_model: AugumentedOnnxModel, diffname, inplace=False):
-        if(inplace):
+    def apply_diff_file(self, augumented_onnx_model: AugumentedOnnxModel, diff_name, diff_order="next_diff", in_place=False):
+        if(in_place):
             new_model = augumented_onnx_model
         else:
             new_model = copy.deepcopy(augumented_onnx_model)
 
-        with open(diffname, "rb") as infile:
+        with open(diff_name, "rb") as infile:
             update_dict = pickle.load(infile)
 
         seen_initializers = set()
@@ -154,12 +158,12 @@ class DiffTools:
                     f"An initializer can appear at most once in a modeldiff file: {initializer_name} appears multiple times."
                 )
 
-            new_model.update_initializer(initializer_name, update_type, delta)
+            new_model.update_initializer(initializer_name, update_type, delta, diff_order)
 
         return new_model
 
 
-    def checkout(self, checkpoint_file):
+    def checkout(self, checkpoint_file, direction="auto"):
         """
         Sequentially applies diffs in order till checkpoint is reached. Order of diffs determined by diff_index of diff_file
         :param diff_file:
@@ -168,7 +172,6 @@ class DiffTools:
         checkpoint_file = os.path.join(constants.COMMITS_FOLDER, checkpoint_file)
         with open(checkpoint_file, "rb") as infile:
             checkpoint_diff = pickle.load(infile)
-        checkpoint_diff_index = checkpoint_diff['diff_index']
 
         with open(self.commit_head_info.global_head, "rb") as infile:
             global_head_diff = pickle.load(infile)
@@ -176,27 +179,44 @@ class DiffTools:
         with open(self.commit_head_info.global_tail, "rb") as infile:
             global_tail_diff = pickle.load(infile)
 
-        if((global_head_diff['diff_index'] - checkpoint_diff_index) <= (checkpoint_diff_index - global_tail_diff['diff_index'])):
-            pointer_direction = "prev_diff"
-            model = AugumentedOnnxModel(self.commit_head_info.global_head_model_path)
-            # Chain starts from diff before global head diff
-            with open(global_head_diff[pointer_direction], "rb") as infile:
-                diff = pickle.load(infile)
-        else:
-            pointer_direction = "next_diff"
-            model = AugumentedOnnxModel(self.commit_head_info.global_tail_model_path)
-            diff = global_tail_diff
+        model, diff, diff_order = self.checkout_init(checkpoint_diff, global_head_diff, global_tail_diff, direction)
 
         # Iterate and update until prior to the checkpoint file using apply_diff_file
         while(diff["current_diff"] != checkpoint_file):
-            model = self.apply_diff_file(model, diff["current_diff"])
-            with open(diff[pointer_direction], "rb") as infile:
+            print(f"Applying diff -: {diff['current_diff']}")
+            model = self.apply_diff_file(model, diff["current_diff"], diff_order=diff_order, in_place=True)
+            with open(diff[diff_order], "rb") as infile:
                 diff = pickle.load(infile)
 
-        # Apply checkpoint diff file
-        model = self.apply_diff_file(model, diff["current_diff"])
+        # Apply checkpoint diff file if moving forwards. If moving backwards, we stop before undoing changes in checkpoint file
+        if(diff_order == "next_diff"):
+            print(f"Applying diff -: {diff['current_diff']}")
+            model = self.apply_diff_file(model, diff["current_diff"], diff_order=diff_order, in_place=True)
 
         return model
+
+
+    def checkout_init(self, checkpoint_diff, global_head_diff, global_tail_diff, direction):
+        backward_diff_count = global_head_diff['diff_index'] - checkpoint_diff['diff_index']
+        forward_diff_count = checkpoint_diff['diff_index'] - global_tail_diff['diff_index']
+        print(f"direction = {direction}")
+        if(direction == "auto"):
+            print(f"forward_diff_count = {forward_diff_count}")
+            print(f"backward_diff_count = {backward_diff_count}")
+
+        if (direction == "backward" or (direction == "auto" and backward_diff_count <= forward_diff_count)):
+            diff_order = "previous_diff"
+            model = AugumentedOnnxModel(self.commit_head_info.global_head_model_path)
+            diff = global_head_diff
+        elif(direction == "forward" or (direction == "auto" and backward_diff_count > forward_diff_count)):
+            diff_order = "next_diff"
+            model = AugumentedOnnxModel(self.commit_head_info.global_tail_model_path)
+            diff = global_tail_diff
+        else:
+            raise ValueError(f"direction should be one of (forward, backward, auto). Value passed -: {direction}")
+
+        print(f"Diff Order -: {diff_order}, Initial diff -: {diff['current_diff']}")
+        return model, diff, diff_order
 
 
     def print_human_readable_diff(self, diff_hash):
